requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.0
Deprecated==1.2.18
Flask==3.1.2
frozendict==2.4.6
frozenlist==1.8.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.20.0
MarkupSafe==3.0.3
multidict==6.7.0
multitasking==0.0.11
numpy==2.2.3
packaging==25.0
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
propcache==0.4.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scratchapi==1.4.1
scratchattach==2.1.9
scratchconnect==5.0
SimpleWebSocketServer==0.1.2
six==1.17.0
soupsieve==2.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
yfinance==0.2.54
flask
line-bot-sdk
Pillow
